chore: clear all clippy warnings across the codebase
- cargo clippy --fix for mechanical lints (redundant closures, derive
  Default, is_none_or, saturating_mul, unnecessary returns/loops)
- manual fixes: merge NaN-rejecting size guards, slice instead of Vec
  params, struct-init defaults in pdf_import, range-loop to iter_mut
- flatten_tree: drop unused layer_idx parameter (and its call sites)
- allow(too_many_arguments) on brush map_point with a rationale comment
- keep rustfmt delta non-positive: 982 pre-existing diffs -> 968

cargo clippy --all-targets: 0 warnings (only the third-party `block
v0.1.6` future-incompat note remains). cargo test: 451 passed, 0 failed.

+++ src/core/brush.rs
@@ -215,6 +215,9 @@ fn frame_at(
 /// Map artwork-space point (x along length, y across) onto the spine.
 /// `x0/x1` bound the artwork's length axis; `y_center` is its across-axis
 /// origin; `y_scale` scales across (stroke-width fit for art brushes).
+// A pure math helper: the spine-frame inputs are natural as scalars, and
+// bundling them into a struct would obscure the call sites more than help.
+#[allow(clippy::too_many_arguments)]
 fn map_point(
     x: f64,
     y: f64,
@@ -314,7 +317,7 @@ pub fn apply_brush_to_polyline(
                 }
             }
             let mut merged = PathData::new();
-            for (band, _) in apply_bristle(&spine_only, *count, *scatter, *size, *opacity as f64) {
+            for (band, _) in apply_bristle(&spine_only, *count, *scatter, *size, *opacity) {
                 merged.elements.extend(band.elements);
             }
             if merged.elements.is_empty() {
@@ -444,7 +447,7 @@ pub fn apply_bristle(
     let count = count.clamp(1, 64);
     let size = size.max(0.5);
     let scatter = scatter.clamp(0.0, 1.0);
-    let opacity = (opacity as f64).clamp(0.0, 1.0);
+    let opacity = opacity.clamp(0.0, 1.0);
     let mut out = Vec::new();
     for (run_idx, (flat, closed)) in spine_runs(spine).iter().enumerate() {
         if flat.len() < 2 {
